+++ tests/hash_test.c
@@ -16,7 +16,7 @@ void f_test(int tam, unsigned int (*hash_func)(int, int, int))
 {  
   Hash *tabela = criar_tabela();
   srand(time(NULL));
-  unsigned long long valor;
+  ULL valor;
   for (int i = 0; i < tam; i++) {
     valor = rand() % TEST_MAX_CHAVES;
     inserir_no(tabela, valor, tam, hash_func);
